Abstract out Motion
